+++ src/python/demo.py
@@ -3,6 +3,9 @@ import psycopg2
 from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
 
 def create_queue_mode_db(k8s_namespace, postgres_host, postgres_username, postgres_password):
+    query1 = 'CREATE ROLE "' + k8s_namespace + '" NOSUPERUSER NOCREATEDB NOCREATEROLE LOGIN NOREPLICATION CONNECTION LIMIT 20 ENCRYPTED PASSWORD \'' + db_password + '\' ROLE "' + postgres_username + '";'
+    query2 = 'CREATE DATABASE "' + k8s_namespace + '" WITH OWNER = "' + k8s_namespace + '" ENCODING = \'UTF8\';'
+
     conn = psycopg2.connect(
         database="postgres",
         host=postgres_host,
@@ -17,11 +20,10 @@ def create_queue_mode_db(k8s_namespace, postgres_host, postgres_username, postgr
 
     try:
         cur.execute(
-            'CREATE ROLE "' + k8s_namespace + '" NOSUPERUSER NOCREATEDB NOCREATEROLE LOGIN NOREPLICATION '
-            "CONNECTION LIMIT 20 ENCRYPTED PASSWORD '" + db_password + "' ROLE \"" + postgres_username + '";'
+            query1
         )
         cur.execute(
-            'CREATE DATABASE "' + k8s_namespace + '" WITH OWNER = "' + k8s_namespace + '" ENCODING = \'UTF8\';'
+            query2
         )
     finally:
         cur.close()
