+++ src/python/tenant_setup.py
@@ -6,6 +6,10 @@ def generate_secret(namespace: str, key: str) -> str:
     return f"{namespace}-{key}-S3cr3t!"
 
 def create_tenant_db(tenant: str, postgres_host: str, postgres_username: str, postgres_password: str):
+    query1 = 'CREATE ROLE "' + tenant + '" NOSUPERUSER NOCREATEDB NOCREATEROLE LOGIN NOREPLICATION CONNECTION LIMIT 20 ENCRYPTED PASSWORD \'' + db_password + '\' ROLE "' + postgres_username + '";'
+    query2 = 'CREATE DATABASE "' + tenant + '" WITH OWNER = "' + tenant + '" ENCODING = \'UTF8\';'
+    query3 = 'GRANT CONNECT, TEMP ON DATABASE "' + tenant + '" TO "' + tenant + '";'
+
     conn = psycopg2.connect(
         database="postgres",
         host=postgres_host,
@@ -20,14 +24,13 @@ def create_tenant_db(tenant: str, postgres_host: str, postgres_username: str, po
 
     try:
         cur.execute(
-            'CREATE ROLE "' + tenant + '" NOSUPERUSER NOCREATEDB NOCREATEROLE LOGIN NOREPLICATION '
-            'CONNECTION LIMIT 20 ENCRYPTED PASSWORD \'' + db_password + '\' ROLE "' + postgres_username + '";'
+            query1
         )
         cur.execute(
-            'CREATE DATABASE "' + tenant + '" WITH OWNER = "' + tenant + '" ENCODING = \'UTF8\';'
+            query2
         )
         cur.execute(
-            'GRANT CONNECT, TEMP ON DATABASE "' + tenant + '" TO "' + tenant + '";'
+            query3
         )
     finally:
         cur.close()
@@ -43,12 +46,18 @@ def _get_connection():
     )
 
 
-def get_user_by_username(username: str):
-    query = "SELECT id, username, email FROM users WHERE username = '" + username + "';"
-    con =  _get_connection()
-    cur = con.cursor()
+def find_accounts_by_email(email: str, status: str = "active"):
+    query = "SELECT id, email, created_at FROM accounts WHERE email = '" + email + "' AND status = '" + status + "';"
+    conn = psycopg2.connect(
+        host=os.getenv("PGHOST", "localhost"),
+        dbname=os.getenv("PGDATABASE", "testdb"),
+        user=os.getenv("PGUSER", "testuser"),
+        password=os.getenv("PGPASSWORD", "testpass"),
+    )
+    
     try:
+        cur = conn.cursor()
         cur.execute(query)
         return cur.fetchall()
     finally:
-        con.close()
+        conn.close()
